Added level of measurement error at follow up
We added another level at 150 cm/s to just have at least 3 categories to compare and see what the trend is.

I also added the baseline true value to the calibration model for regression calibration method. That would give the imputation and RC methods the same information.

However, I wonder if that is appropriate when we are then looking at the *difference* as the exposure?