Add different sample sizes
This commit has quite a few changes. In trying to use smaller sample sizes, it became apparent that we were hitting a lower bound on how many participants could reasonably be used in the calibration study. For example, with n = 100 and using 2 people in the calibration study, you can't have 5 different parameters in your regression model (duh!).

So, we ended up limiting to n = 500 as the smallest study total sample size and adjusted the smallest percent of those being sampled to 5%.

We have also adjusted from sampling the participants using a binomial with p of some value (e.g., 0.05) to actual sampling a specified number of participants. For example, when n = 500 and p = 0.05, we will always sample 25 participants.