Swtch to tidyverse from data.table
I think I made some huge mistakes with plotting the results for the right method when I was using data.table. I've gone back to tidyverse just because it's what I know and I am confident that it is right. It's possible that some of the results were mislabeled in previous versions of the manuscript.